new feat: Base architecture for the run app func
Started writing the run_app function that will take in an image path, and depending on if the image is a dog or a human, will output either the predicted dog breed or the breed the human most resembles.  Current implementation just runs the dog_detect and human_detect functions defined earlier.